requirements.txt
Flask==2.2.5
GitPython==3.1.43
lizard==1.17.10
flask-cors==4.0.0
